update handshake response handling
