+++ dot/network/notifications.go
@@ -136,16 +136,9 @@ func (s *Service) createNotificationsMessageHandler(info *notificationsProtocol,
 				}
 				logger.Trace("receiver: sent handshake", "sub-protocol", info.subProtocol, "peer", peer)
 
-				if handshakeResp != nil {
-					logger.Debug("sending message to peer in response to handshake", "peer", peer, "msg", handshakeResp)
-					s.syncing[peer] = struct{}{}
-					err = s.host.send(peer, syncID, handshakeResp) // TODO: return protocol from handshakeValidator, or put protocol into Message interface?
-					if err != nil {
-						logger.Error("failed to send response to peer's handshake", "sub-protocol", info.subProtocol, "peer", peer, "error", err)
-						return err
-					}
-
-					go s.handleSyncStream(s.host.getStream(peer, syncID))
+				err = s.handleHandshakeResponse(peer, handshakeResp)
+				if err != nil {
+					return err
 				}
 			}
 
@@ -163,13 +156,9 @@ func (s *Service) createNotificationsMessageHandler(info *notificationsProtocol,
 				info.handshakeData[peer].received = true
 				logger.Trace("sender: validated handshake", "sub-protocol", info.subProtocol, "peer", peer)
 
-				if handshakeResp != nil {
-					logger.Debug("sending message to peer in response to handshake", "peer", peer, "msg", handshakeResp)
-					err = s.host.send(peer, syncID, handshakeResp) // TODO: fix protocol ID
-					if err != nil {
-						logger.Error("failed to send response to peer's handshake", "sub-protocol", info.subProtocol, "peer", peer, "error", err)
-						return err
-					}
+				err = s.handleHandshakeResponse(peer, handshakeResp)
+				if err != nil {
+					return err
 				}
 			} else if hsData.received {
 				return nil
@@ -204,6 +193,29 @@ func (s *Service) createNotificationsMessageHandler(info *notificationsProtocol,
 
 		return nil
 	}
+}
+
+// handleHandshakeResponse handles a potential response to a peer's handshake.
+// this is currently only used for syncing (ie. if we received a BlockAnnounceHandhsake where the peer has a higher chain head
+// than us, we want to begin syncing with them.)
+func (s *Service) handleHandshakeResponse(peer peer.ID, msg Message) error {
+	if msg == nil {
+		return nil
+	}
+
+	logger.Debug("sending message to peer in response to handshake", "peer", peer, "msg", msg)
+	switch msg.SubProtocol() {
+	case syncID:
+		err := s.beginSyncing(peer, msg)
+		if err != nil {
+			logger.Error("failed to send response to peer's handshake", "sub-protocol", syncID, "peer", peer, "error", err)
+			return err
+		}
+	default:
+		return errors.New("unsupported handshake response")
+	}
+
+	return nil
 }
 
 // broadcastExcluding sends a message to each connected peer except the given peer
